modularized code files/structure (IMPORTS TO BE FIXED)
