+++ src/structs/info_address_port_pair.rs
@@ -2,8 +2,9 @@
 //! to keep track of statistics about the sniffed traffic.
 
 use std::fmt;
+use crate::enums::traffic_type::TrafficType;
 
-use crate::structs::address_port_pair::TrafficType;
+use crate::utility::get_formatted_strings::get_formatted_bytes_string;
 
 /// Struct useful to format the output report file and to keep track of statistics about the sniffed traffic.
 ///
@@ -56,154 +57,4 @@ impl fmt::Display for InfoAddressPortPair {
                    self.initial_timestamp, self.final_timestamp, " ".repeat(40))
         }
     }
-}
-
-
-pub fn get_formatted_bytes_string(bytes: u128) -> String {
-    let mut multiple_transmitted = "".to_string();
-    let mut n = bytes as f32;
-
-    match bytes {
-        0..=999 => {}
-        1_000..=999_999 => {
-            n /= 1000_f32;
-            multiple_transmitted.push('k');
-        } // kilo
-        1_000_000..=999_999_999 => {
-            n /= 1_000_000_f32;
-            multiple_transmitted.push('M');
-        } // mega
-        1_000_000_000..=999_999_999_999 => {
-            n /= 1_000_000_000_f32;
-            multiple_transmitted.push('G');
-        } // giga
-        _ => {
-            n /= 1_000_000_000_000_f32;
-            multiple_transmitted.push('T');
-        } // tera
-    }
-
-    if !multiple_transmitted.is_empty() { // with multiple
-        format!("{:.1} {}B", n, multiple_transmitted)
-    } else { // no multiple
-        format!("{}  B", n)
-    }
-}
-
-
-/// Enum representing the possible observed values of transport layer protocol.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum TransProtocol {
-    /// Transmission Control Protocol
-    TCP,
-    /// User Datagram Protocol
-    UDP,
-    /// Not identified
-    Other,
-}
-
-
-impl fmt::Display for TransProtocol {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-
-
-/// Enum representing the possible observed values of application layer protocol.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum AppProtocol {
-    /// File Transfer Protocol
-    FTP,
-    /// Secure Shell
-    SSH,
-    /// Telnet
-    Telnet,
-    /// Simple Mail Transfer Protocol
-    SMTP,
-    /// Terminal Access Controller Access-Control System
-    TACACS,
-    /// Domain Name System
-    DNS,
-    /// Dynamic Host Configuration Protocol
-    DHCP,
-    /// Trivial File Transfer Protocol
-    TFTP,
-    /// Hypertext Transfer Protocol
-    HTTP,
-    /// Post Office Protocol
-    POP,
-    /// Network Time Protocol
-    NTP,
-    /// NetBIOS
-    NetBIOS,
-    /// Post Office Protocol 3 over TLS/SSL
-    POP3S,
-    /// Internet Message Access Protocol
-    IMAP,
-    /// Simple Network Management Protocol
-    SNMP,
-    /// Border Gateway Protocol
-    BGP,
-    /// Lightweight Directory Access Protocol
-    LDAP,
-    ///Hypertext Transfer Protocol over TLS/SSL
-    HTTPS,
-    /// Lightweight Directory Access Protocol over TLS/SSL
-    LDAPS,
-    /// File Transfer Protocol over TLS/SSL
-    FTPS,
-    /// Multicast DNS
-    #[allow(non_camel_case_types)]
-    mDNS,
-    ///Internet Message Access Protocol over TLS/SSL
-    IMAPS,
-    /// Simple Service Discovery Protocol
-    SSDP,
-    /// Extensible Messaging and Presence Protocol |
-    XMPP,
-    /// not identified
-    Other,
-}
-
-
-impl fmt::Display for AppProtocol {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        if self.eq(&AppProtocol::Other) {
-            write!(f, "All protocols")
-        } else {
-            write!(f, "{:?}", self)
-        }
-    }
-}
-
-/// Defines a constant to be used in the picklist in gui initial page
-impl AppProtocol {
-    pub(crate) const ALL: [AppProtocol; 25] = [
-        AppProtocol::Other,
-        AppProtocol::BGP,
-        AppProtocol::DHCP,
-        AppProtocol::DNS,
-        AppProtocol::FTP,
-        AppProtocol::FTPS,
-        AppProtocol::HTTP,
-        AppProtocol::HTTPS,
-        AppProtocol::IMAP,
-        AppProtocol::IMAPS,
-        AppProtocol::LDAP,
-        AppProtocol::LDAPS,
-        AppProtocol::mDNS,
-        AppProtocol::NetBIOS,
-        AppProtocol::NTP,
-        AppProtocol::POP,
-        AppProtocol::POP3S,
-        AppProtocol::SMTP,
-        AppProtocol::SNMP,
-        AppProtocol::SSDP,
-        AppProtocol::SSH,
-        AppProtocol::TACACS,
-        AppProtocol::Telnet,
-        AppProtocol::TFTP,
-        AppProtocol::XMPP,
-    ];
 }
