+++ src/structs/mod.rs
@@ -3,3 +3,5 @@ pub mod info_address_port_pair;
 pub mod info_traffic;
 pub mod runtime_data;
 pub mod traffic_chart;
+pub mod filters;
+pub mod sniffer;
